requirements.txt
faiss-cpu==1.7.4
fastapi==0.104.0
numpy==1.26.1
pandas==2.1.2
uvicorn==0.23.2
joblib==1.3.2